STM32 lwIP addition and CMake support
This patch adds CMake support to RTEMS lwIP.
It also improves the architecture to make integration
of new BSPs easier.

https://github.com/rmspacefish/rtems-stm32-lwip is a self-contained
repository to test the lwIP integration.

The repository now requires the user to supply the lwipopts.h
file by setting the CMake variable LWIP_CONFIG_PATH in
the upper CMakeLists.txt file. The user also needs to
set RTEMS_BSP. THe RTEMS lwIP integration will then
select the correct port based on the BSP information.

The STM32 port includes the LAN driver (lan8742.c), the
primary port functions in ethernetif.c and some application code
taken from CubeH7, which can serve as a starting point for
implementations and was also tailored to RTEMS.

The common source file were refactored to be more generic and independent
of the used BSP.