CMake support for the STM32 lwIP addition
This patch adds full CMake support for the lwIP support.
It can be easily extended to other architectures as well.

More information can also be found in the README, which also specifies
how to build with CMake. The waf build system still needs to be adapted
to perform the same function.